feat: gpu 삭제 관련 테스트코드 추가

+++ back-end/src/main/java/admin/gpu/application/GpuService.java
@@ -12,18 +12,20 @@ import admin.gpu.dto.GpuServerNameUpdateRequest;
 import admin.gpu.dto.GpuServerRequest;
 import admin.gpu.dto.GpuServerResponse;
 import admin.gpu.dto.GpuServerResponses;
+import admin.gpu.exception.GpuServiceException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class GpuService {
+
     private LabRepository labRepository;
     private GpuServerRepository gpuServerRepository;
     private GpuBoardRepository gpuBoardRepository;
     private DeleteHistoryRepository deleteHistoryRepository;
 
     public GpuService(LabRepository labRepository, GpuServerRepository gpuServerRepository,
-                      GpuBoardRepository gpuBoardRepository, DeleteHistoryRepository deleteHistoryRepository) {
+        GpuBoardRepository gpuBoardRepository, DeleteHistoryRepository deleteHistoryRepository) {
         this.labRepository = labRepository;
         this.gpuServerRepository = gpuServerRepository;
         this.gpuBoardRepository = gpuBoardRepository;
@@ -42,24 +44,35 @@ public class GpuService {
 
     @Transactional
     public void updateGpuServer(GpuServerNameUpdateRequest gpuServerNameUpdateRequest, Long labId,
-                                Long gpuServerId) {
+        Long gpuServerId) {
         GpuServer gpuServer = gpuServerRepository.findById(gpuServerId).get();
         gpuServer.setName(gpuServerNameUpdateRequest.getName());
     }
 
     @Transactional
     public void delete(Long labId, Long gpuId) {
+        if (isNotExistGpuServer(gpuId)) {
+            throw new GpuServiceException("Gpu Id가 존재하지 않습니다.");
+        }
         GpuServer gpuServer = gpuServerRepository.findById(gpuId).get();
+
+        if (gpuServer.getDeleted()) {
+            throw new GpuServiceException("Gpu 가 이미 삭제된 상태입니다");
+        }
         gpuServer.setDeleted(true);
         deleteHistoryRepository.save(new DeleteHistory(gpuServer));
+    }
+
+    private boolean isNotExistGpuServer(Long gpuId) {
+        return !gpuServerRepository.existsById(gpuId);
     }
 
     @Transactional
     public Long saveGpuServer(GpuServerRequest gpuServerRequest, Long labId) {
         Lab lab = labRepository.findById(labId).get();
         GpuServer gpuServer = new GpuServer(gpuServerRequest.getServerName(),
-                gpuServerRequest.getMemorySize(),
-                gpuServerRequest.getMemorySize(), lab);
+            gpuServerRequest.getMemorySize(),
+            gpuServerRequest.getMemorySize(), lab);
         gpuServerRepository.save(gpuServer);
         return gpuServer.getId();
     }
